Add CalculatedPay attribute with PayCommand
Implemented CalculatedPay as an element in Salary, and can be accessed with its get method. Since Salary is immutable, CalculatedPay is set to be final.
With this implementation, setting the calculatedPay of a person is done by creating a new person with the overloaded constructor method and replacing the person to be paid with the new person.
This is also done for the PayCommand implementation, which takes in the index and sets the calculatedPay of a person to 0.